requirements.txt
aiogram==3.6.0
SQLAlchemy==2.0.32
pydantic==2.7.1
python-dotenv==1.0.1
aiosqlite==0.20.0